* 'master-from-1.1' of scm.dev.nokia.troll.no:qtmobility/qtmobility-integration: (158 commits)
  Fixed compile of qgeoserviceprovider tests.
  maemo5: fix QTMOBILITY-891: mkCal async engine implementation is used
  QMediaPlayer: remove the position property watch after playback is stopped.
  Fix the order of filters inserted by QGallery(*)Filter.prepend().
  Gstreamer player backend: Automatically detect if the colorspace element is necessary.
  Forward focusZonesChanged() signal from QCameraFocusControl to QCameraFocus
  Need to mention dependency.
  Mention MatchContains is not supported on Symbian
  really set the screen blanking.
  Stabilization of auto-test
  doc fix: volume is a qreal
  Advertise if a media service supports stream playback.
  Don't print unnecesary warnings.
  offset function of landmark fetch requests not adequately documented
  Mention dependency on QtSparql Tracker extensions
  maemo6: Landmarks, Removed country code handling from qsparql plugin
  maemo6: Landmarks, modified dataChanged() signalling for large operations
  MOBILITY-2173 Memory leak in CServiceProviderServerSession
  Fix typo
  Better documentation for offset functions.
  ...

+++ tests/auto/qlandmarkfilehandler_gpx/tst_qlandmarkfilehandler_gpx.cpp
@@ -49,10 +49,12 @@
 #include <QBuffer>
 #include <QDebug>
 #include <QThread>
+#include <QScopedPointer>
 
 #define private public
 #include <qlandmarkmanager.h>
 #include "qlandmarkfilehandler_gpx_p.h"
+#include "qlandmarkmanagerdataholder.h"
 
 
 QTM_USE_NAMESPACE
@@ -65,7 +67,7 @@ public:
 
 protected:
     void run() {
-        msleep(50);
+        msleep(10);
         m_cancel = true;
     }
 };
@@ -81,32 +83,42 @@ private:
     QLandmarkManager *m_manager;
     QLandmarkFileHandlerGpx *m_handler;
     QString m_exportFile;
+    QScopedPointer<QLandmarkManagerDataHolder> managerDataHolder;
+
+    void clean() {
+        QList<QLandmarkId> lmIds = m_manager->landmarkIds();
+        for(int i=0; i < lmIds.count(); ++i) {
+            QVERIFY(m_manager->removeLandmark(lmIds.at(i)));
+        }
+
+        QList<QLandmarkCategoryId> catIds = m_manager->categoryIds();
+        for (int i=0; i < catIds.count(); ++i) {
+            if (!m_manager->isReadOnly(catIds.at(i)))
+                QVERIFY(m_manager->removeCategory(catIds.at(i)));
+        }
+        QTest::qWait(20);//try ensure notifications for these
+                         //deletions are made prior to each test function.
+    }
 
 private slots:
 
     void init() {
-        QMap<QString, QString> map;
-        map["filename"] = "test.db";
-        m_exportFile = "exportfile";
-        m_manager = new QLandmarkManager("com.nokia.qt.landmarks.engines.sqlite", map);
-
+        m_manager = new QLandmarkManager;
         m_handler = new QLandmarkFileHandlerGpx;
+        clean();
     }
 
     void cleanup() {
+        clean();
         delete m_handler;
         delete m_manager;
 
-        QFile file("test.db");
-        file.remove();
-        QFile::remove(m_exportFile);
+        if (QFile::exists(m_exportFile))
+            QFile::remove(m_exportFile);
     }
 
-    void cleanupTestCase() {
-        QFile file("test.db");
-        file.remove();
-        QFile::remove(m_exportFile);
-    }
+    void initTestCase();
+    void cleanupTestCase();
 
     void fileImport() {
         QFETCH(QString, fileIn);
@@ -185,7 +197,7 @@ private slots:
         QLandmarkFileHandlerGpx handler(&(cancelThread.m_cancel));
         QLandmark lm;
         QList<QLandmark> lms;
-        for (int i=0; i < 50000; ++i) {
+        for (int i=0; i < 100000; ++i) {
             lm.setName(QString("LM%1").arg(0));
             lms.append(lm);
         }
@@ -193,7 +205,11 @@ private slots:
         handler.setWaypoints(lms);
         cancelThread.start();
         QFile file(m_exportFile);
-        QVERIFY(!handler.exportData(&file));
+        bool result = handler.exportData(&file);
+        if (result)
+            cancelThread.wait();
+
+        QVERIFY(!result);
         QCOMPARE(handler.error(), QLandmarkManager::CancelError);
         cancelThread.wait();
     }
@@ -500,6 +516,17 @@ private:
         << r;
     }
 };
+
+void tst_QLandmarkFileHandler_Gpx::initTestCase()
+{
+    managerDataHolder.reset(new QLandmarkManagerDataHolder());
+}
+
+void tst_QLandmarkFileHandler_Gpx::cleanupTestCase() {
+    managerDataHolder.reset(0);
+    if (QFile::exists(m_exportFile))
+        QFile::remove(m_exportFile);
+}
 
 QTEST_MAIN(tst_QLandmarkFileHandler_Gpx)
 #include "tst_qlandmarkfilehandler_gpx.moc"
